+++ Source/Scripts/Spells/Shaman.cs
@@ -633,7 +633,7 @@ namespace Scripts.Spells.Shaman
             Player player = GetCaster().ToPlayer();
             byte slot = EquipmentSlot.MainHand;
 
-            if (player.GetPrimarySpecialization() == (uint)TalentSpecialization.ShamanEnhancement)
+            if (player.GetPrimarySpecialization() == TalentSpecialization.ShamanEnhancement)
                 slot = EquipmentSlot.OffHand;
 
             Item targetItem = player.GetItemByPos(InventorySlots.Bag0, slot);
